Superman Media Dataset

A dataset containing information about Superman media appearances across film, television, and animation, including actor heights, critical reception, and audience engagement metrics.

Format

A tibble with 11 rows and 21 variables:

num

Numeric identifier for each media entry

media

Media identifier

year

Year of release

type

Media type (1 = Film, 2 = TV, 3 = Serial)

clark_height

Actor height in meters

lois_height

Actress height in meters

rt_critics_score

Rotten Tomatoes critics score (0-100)

rt_critic_count

Number of RT critic reviews

rt_audience_score

Rotten Tomatoes audience score (0-100)

rt_audience_count

Number of RT audience ratings

lbd_likes

Letterboxd likes count

lbd_scores

Letterboxd average score (0-5)

clark_height_in

Actor height in inches

lois_height_in

Actress height in inches

clark_grp

Height group (1 = under 6ft, 2 = 6ft+)

height_diff

Height difference in inches (Clark - Lois)

height_gap

Height gap category (1 = <6in, 2 = 6-8in, 3 = >8in)

tomatometer

Tomatometer (1 = Rotten, 2 = Fresh)

rt_avg

Average of critics and audience scores

rt_diff

Weighted difference between critics and audience

popular

Popularity category (1 = low, 2 = mid, 3 = high)

Source

Rotten Tomatoes, Letterboxd, and public biographical data.

Examples

data(superman)
head(superman)
#> # A tibble: 6 × 21
#>     num media  year  type clark_height lois_height rt_critics_score
#>   <int> <dbl> <dbl> <dbl>        <dbl>       <dbl>            <dbl>
#> 1     1     1  2025     1         1.93        1.6                83
#> 2     2     2  1978     1         1.93        1.72               88
#> 3     3     3  2001     2         1.9         1.71               78
#> 4     4     4  2006     1         1.89        1.65               72
#> 5     5     5  1951     3         1.86        1.63               NA
#> 6     6     6  2013     1         1.85        1.63               57
#> # ℹ 14 more variables: rt_critic_count <dbl>, rt_audience_score <dbl>,
#> #   rt_audience_count <dbl>, ldb_likes <dbl>, ldb_scores <dbl>,
#> #   clark_height_in <dbl>, lois_height_in <dbl>, clark_grp <dbl>,
#> #   height_diff <dbl>, height_gap <dbl>, tomatometer <dbl>, rt_avg <dbl>,
#> #   rt_diff <dbl>, popular <dbl>
mean(superman$clark_height_in, na.rm = TRUE)
#> [1] 73.37136